+++ lib/components/z-inputs/z-input-generic.dart
@@ -5,6 +5,7 @@ import 'package:z_components/styles/main-style.dart';
 
 class ZInputGeneric extends StatefulWidget {
   final Key key;
+  final ThemeData themeData;
   String hintText;
   String textMask;
   String titulo;
@@ -25,6 +26,7 @@ class ZInputGeneric extends StatefulWidget {
       this.obscureText: false,
       this.tipoTeclado: TextInputType.text,
       this.comMascara: false,
+      @required this.themeData,
       @required this.titulo,
       this.textMask,
       @required this.inputPadraoFocus,
@@ -60,8 +62,9 @@ class _ZInputGenericState extends State<ZInputGeneric> {
                   child: new Container(
                     padding: const EdgeInsets.only(
                         top: 12.0, bottom: 12.0, left: 16.0),
-                    child: new Text("${widget.titulo}:",
-                        style: MainStyle.get(context).mainStyleTextBaseLine),
+                    child: new Text("${widget.titulo}",
+                        style: widget.themeData.textTheme.bodyText1
+                            .copyWith(color: Color(0XFF999999))),
                   ),
                 ),
                 new Expanded(
@@ -78,8 +81,7 @@ class _ZInputGenericState extends State<ZInputGeneric> {
                             obscureText: widget.obscureText,
                             controller: widget.controllerInputPadrao,
                             cursorColor: Color(0xFF801F92),
-                            style: MainStyle.get(context)
-                                .mainStyleTextBaseLineInput,
+                            style: widget.themeData.textTheme.bodyText1,
                             inputFormatters: [
                               MaskedTextInputFormatterShifter(
                                   maskONE: widget.textMask,
@@ -94,8 +96,8 @@ class _ZInputGenericState extends State<ZInputGeneric> {
                             },
                             decoration: InputDecoration(
                               hintText: widget.hintText,
-                              hintStyle: MainStyle.get(context)
-                                  .mainStyleTextBaseLineHint,
+                              hintStyle: widget.themeData.textTheme.bodyText1
+                                  .copyWith(color: Color(0xFFC7C7C7)),
                               enabledBorder: UnderlineInputBorder(
                                 borderSide:
                                     BorderSide(color: Color(0xFFf0f0f0)),
@@ -119,8 +121,7 @@ class _ZInputGenericState extends State<ZInputGeneric> {
                             obscureText: widget.obscureText,
                             controller: widget.controllerInputPadrao,
                             cursorColor: Color(0xFF801F92),
-                            style: MainStyle.get(context)
-                                .mainStyleTextBaseLineInput,
+                            style: widget.themeData.textTheme.bodyText1,
                             onSubmitted: (text) {
                               widget.inputPadraoFocus.unfocus();
                               if (widget.proximoFocus != null) {
@@ -130,8 +131,8 @@ class _ZInputGenericState extends State<ZInputGeneric> {
                             },
                             decoration: InputDecoration(
                               hintText: widget.hintText,
-                              hintStyle: MainStyle.get(context)
-                                  .mainStyleTextBaseLineHint,
+                              hintStyle: widget.themeData.textTheme.bodyText1
+                                  .copyWith(color: Color(0xFFC7C7C7)),
                               enabledBorder: UnderlineInputBorder(
                                 borderSide:
                                     BorderSide(color: Color(0xFFf0f0f0)),