+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 0.9.30
+
+* Update ZSequenciaCalendario
+
+## 0.9.29+3
+
+* Update ZInput com theme
+
+## 0.9.29+2
+
+* Update ZInput com theme
+
+## 0.9.29+1
+
+* Update ZInput com theme
+
+## 0.9.29
+
+* Update ZInput recebendo theme
+
+## 0.9.28+2
+
+* Update Z Sequencia Calendario 
+
 ## 0.9.28+1
 
 * Update Z Sequencia Calendario